Moved InputField into directory with css
