+++ client/src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as SampleComponent } from "./SampleComponent";
-export * from "./InputField";
+export * from "./InputField/InputField";
 export * from "./ProductContainer";
 export * from "./SampleComponent";
 export * from "./SearchParamForm";
